Add disassembler and reformat to store in big-endian
Perform little-endian -> big-endian conversion once on read of program
so we can use native numbers.

The main reason I did this was because there was a subtle bug causing a
room to not print the full description, but this error was silent!
Unfortunately the text not being printed was the coin puzzle clue
which is required to be able to solve the puzzle.

Refactoring has fixed the issue (though I'm still not sure how).

Disassembler added to help diagnore the above problem (spoiler - it
didn't).